// frontend/src/types/api.ts
export interface Page<T> {
  items: T[];
  page: number;
  limit: number;
  total: number;
  pages: number;
}
// ...
export interface Tag {
  slug: string;
  name: string;
}

export interface TagStat extends Tag {
  posts_count: number;
}
// ...
export type SearchType = 'all' | 'posts' | 'users' | 'tags';